# openapi.yaml
openapi: 3.0.3
info:
  title: API MiniBlog
  description: API REST de un mini blog con autores y posts.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Entorno local
  - url: https://apiminiblog-production.up.railway.app
    description: Producción (Railway)

paths:
  /authors:
    get:
      tags:
        - Autores
      summary: Lista todos los autores
      responses:
        '200':
          description: Lista de autores obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Author'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Autores
      summary: Crea un nuevo autor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: Juan Pérez
                email:
                  type: string
                  example: juan@example.com
                bio:
                  type: string
                  example: Desarrollador Backend
      responses:
        '201':
          description: Autor creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: name o email vacíos, JSON mal formado o email duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /authors/{id}:
    get:
      tags:
        - Autores
      summary: Obtiene un autor por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Autor encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: ID inválido (debe ser un entero positivo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Autores
      summary: Actualiza un autor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
              properties:
                name:
                  type: string
                  example: Juan Pérez
                email:
                  type: string
                  example: juan@example.com
                bio:
                  type: string
                  example: Desarrollador Backend
      responses:
        '200':
          description: Autor actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Author'
        '400':
          description: ID inválido, name o email vacíos, JSON mal formado o email duplicado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Autores
      summary: Elimina un autor (también elimina sus posts por CASCADE)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Autor eliminado correctamente (sin contenido)
        '400':
          description: ID inválido (debe ser un entero positivo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Autor no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /posts:
    get:
      tags:
        - Posts
      summary: Lista todos los posts
      responses:
        '200':
          description: Lista de posts obtenida correctamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - Posts
      summary: Crea un nuevo post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - author_id
                - title
                - content
              properties:
                author_id:
                  type: integer
                  example: 1
                title:
                  type: string
                  example: Mi primer post
                content:
                  type: string
                  example: Contenido detallado del post...
                published:
                  type: boolean
                  default: false
                  example: false
      responses:
        '201':
          description: Post creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: Faltan datos obligatorios, JSON mal formado o el autor no existe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /posts/author/{authorId}:
    get:
      tags:
        - Posts
      summary: Obtiene los posts de un autor con el detalle del autor
      parameters:
        - name: authorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Lista de posts del autor con nombre y email del autor (vacía si no tiene posts)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostWithAuthor'
        '400':
          description: ID de autor inválido (debe ser un entero positivo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

  /posts/{id}:
    get:
      tags:
        - Posts
      summary: Obtiene un post por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Post encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: ID inválido (debe ser un entero positivo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - Posts
      summary: Actualiza un post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - author_id
                - title
                - content
              properties:
                author_id:
                  type: integer
                  example: 1
                title:
                  type: string
                  example: Título actualizado
                content:
                  type: string
                  example: Contenido actualizado del post...
                published:
                  type: boolean
                  example: true
      responses:
        '200':
          description: Post actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: ID inválido, datos inválidos o JSON mal formado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - Posts
      summary: Elimina un post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Post eliminado correctamente (sin contenido)
        '400':
          description: ID inválido (debe ser un entero positivo)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Post no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  responses:
    ServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Mensaje que describe el error

    Author:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Juan Pérez
        email:
          type: string
          example: juan@example.com
        bio:
          type: string
          example: Desarrollador Backend
        created_at:
          type: string
          format: date-time
          example: '2026-10-05T23:05:18.187Z'

    Post:
      type: object
      properties:
        id:
          type: integer
          example: 1
        author_id:
          type: integer
          example: 1
        title:
          type: string
          example: Introducción a PostgreSQL
        content:
          type: string
          example: En este artículo aprenderás los conceptos básicos...
        published:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: '2026-10-05T23:05:18.187Z'

    PostWithAuthor:
      allOf:
        - $ref: '#/components/schemas/Post'
        - type: object
          properties:
            author_name:
              type: string
              example: Juan Pérez
            author_email:
              type: string
              example: juan@example.com